add runtime setting for streamfunction solver
