+++ veros/core/streamfunction/solve_stream.py
@@ -5,7 +5,7 @@ with Dirichlet boundary conditions
 used for streamfunction
 """
 
-from . import utilities, solve_poisson_petsc_full
+from . import utilities
 from .. import utilities as mainutils
 from ... import veros_method, runtime_settings as rs
 
@@ -52,13 +52,11 @@ def solve_streamfunction(vs):
     # solve for interior streamfunction
     vs.dpsi[:, :, vs.taup1] = 2 * vs.dpsi[:, :, vs.tau] - vs.dpsi[:, :, vs.taum1]
 
-    # dpsi_global = gather(vs, vs.dpsi[..., vs.taup1], ("xt", "yt"))
-    solve_poisson_petsc_full.solve(
+    vs.linear_solver.solve(
         vs,
         forc,
         vs.dpsi[..., vs.taup1]
     )
-    # vs.dpsi[..., vs.taup1] = scatter(vs, dpsi_global, ("xt", "yt"))
 
     mainutils.enforce_boundaries(vs, vs.dpsi[:, :, vs.taup1])
