+++ veros/settings.py
@@ -130,7 +130,6 @@ SETTINGS = OrderedDict([
     ("pyom_compatibility_mode", Setting(False, bool, "Force compatibility to pyOM2 (even reproducing bugs and other quirks). For testing purposes only.")),
     ("diskless_mode", Setting(False, bool, "Suppress all output to disk. Mainly used for testing purposes.")),
     ("default_float_type", Setting("float64", str, "Default type to use for floating point arrays (e.g. ``float32`` or ``float64``).")),
-    ("use_amg_preconditioner", Setting(True, bool, "Use AMG preconditioner in Poisson solver if pyamg is installed."))
 ])
